refactor: separation of responsibilities
